requirements.txt
algoliasearch-django
crispy-bootstrap4
django-allauth
Django==4.1
django-cors-headers==4.3.1
django-crispy-forms
django-debug-toolbar==4.2.0
django-filter==23.3
django-select2==8.1.2
django-simple-menu==2.1.3
django-tables2==2.6.0
dj-database-url==2.1.0
drf-extra-fields==3.7.0
Pillow==10.1.0
python-decouple==3.8
django-allauth
